Fix incorrect curl command in API docs
The docs suggested that you only needed to use your api-base-url in
the curl command, but actually it needs to have /organizations appended
to it in order to work correctly. This makes it consistent with other
commands on this page which explicitly show a url to use.

I've also take the opportunity to reorder the curl arguments so the url
comes last, and use the long version of the option names for clarity.